requirements.txt
mcp-server
requests
pandas 